+++ anetbbs/__init__.py
@@ -1,4 +1,4 @@
 """
 AnetBBS - A Modern Python-based BBS system with Web and Telnet support
 """
-__version__ = '1.0b2.235'
+__version__ = '1.0b2.238'